Feature: budget 테이블 이름 변경 sql 스트립트 추가(budget_by_category)
- Budget.java -> CategoryBudget.java
- BudgetTest.java -> CategoryBudgetTest.java
- CategoryBudget Date 필드 삭제
