+++ src/main/java/jaringobi/domain/budget/CategoryBudget.java
@@ -7,20 +7,17 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
 import jaringobi.domain.BaseTimeEntity;
 import jaringobi.exception.budget.InvalidBudgetException;
-import java.util.Date;
 import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "budget")
+@Table(name = "budget_by_category")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Budget extends BaseTimeEntity {
+public class CategoryBudget extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -32,20 +29,15 @@ public class Budget extends BaseTimeEntity {
     @Column(nullable = false)
     private Long categoryId;
 
-    @Temporal(value = TemporalType.DATE)
-    @Column(nullable = false)
-    private Date budgetMonth;
-
     @Builder
-    public Budget(final Money amount, final Long categoryId, final Date budgetMonth) {
-        validateBudget(amount, categoryId, budgetMonth);
+    public CategoryBudget(final Money amount, final Long categoryId) {
+        validateBudget(amount, categoryId);
         this.amount = amount;
         this.categoryId = categoryId;
-        this.budgetMonth = budgetMonth;
     }
 
-    private void validateBudget(Money amount, Long categoryId, Date budgetMonth) {
-        if (Objects.isNull(amount) || Objects.isNull(budgetMonth) || Objects.isNull(categoryId)) {
+    private void validateBudget(Money amount, Long categoryId) {
+        if (Objects.isNull(amount) || Objects.isNull(categoryId)) {
             throw new InvalidBudgetException();
         }
     }
