+++ src/main/resources/db/migration/V3.1__change_table_name_budget.sql
@@ -1,0 +1,1 @@
+ALTER TABLE budget RENAME TO budget_by_category;
